switched to orb features for stereo matching
first look for features in left image and get as many as possible that are sufficiently far away from existing features, then look for stereo matches.
this puts more importance on the distribution of the features in the image vs the quality of the stereo matches

the search for distant features is iterated with successively smaller distance requirements until the number of requested features is found or a maximum of iterations is reached.
after matching, we still may end up with fewer features than requested, if not every feature was able to be matched (this happens even though many features are already matched with very bad correspondences)